Fixed game loop. More settings added to game.lua script

+++ src/entities/Character.h
@@ -13,7 +13,7 @@ public:
 
 	virtual void Draw() = 0;
 	virtual void Clean() = 0;
-	virtual void Update( float dt ) = 0;
+	virtual void Update( double dt ) = 0;
 
 
 protected:
